Changed power amount of specials from uint to FP
certain specials will cast to a uint for damage purposes

+++ Quantum/quantum_code/quantum.code/Components/SpecialShieldedCharge.cs
@@ -77,7 +77,7 @@ namespace Quantum
 				ChargeStartPos = attackerPosition,
 				ChargeEndPos = targetPosition,
 				ChargeStartTime = f.Time,
-				PowerAmount = special.PowerAmount
+				PowerAmount = (UInt32)special.PowerAmount
 			};
 			
 			QuantumHelpers.LookAt2d(f, e, targetPosition);
